completed auth refactoring and corrected imports. Also removed unnecessary files
